Enable (very) basic compute kernels (#37328)
### Ticket
N/A

### Problem description
Quasar wasn't running any compute kernels

### What's changed
This PR adds compute FW and support to run compute kernels on Neo0
triscs. It only alows to run basic code on hte trisc and doesn't support
any real compute workloads. That work is included on a different branch
that has LLK API changes etc. The goal of this PR is to unblock others
that needs to test features that relate to triscs.

### Checklist

- [x] [![All post-commit
tests](https://github.com/tenstorrent/tt-metal/actions/workflows/all-post-commit-workflows.yaml/badge.svg?branch=ayaacob/compute_fw)](https://github.com/tenstorrent/tt-metal/actions/runs/21767356986)
- [x] [![Blackhole Post
commit](https://github.com/tenstorrent/tt-metal/actions/workflows/blackhole-post-commit.yaml/badge.svg?branch=ayaacob/compute_fw)](https://github.com/tenstorrent/tt-metal/actions/runs/21767228933)
- [x] New/Existing tests provide coverage for changes - added a test
that run a simple compute kernel that adds numbers on RISC-V and prints.

---------

+++ tt_metal/hw/inc/api/remote_circular_buffer.h
@@ -74,7 +74,7 @@ FORCE_INLINE void update_pages_acked(
     volatile tt_l1_ptr uint32_t* pages_acked_ptr =
         reinterpret_cast<volatile tt_l1_ptr uint32_t*>(aligned_pages_acked_addr);
     *pages_acked_ptr += aligned_page_adjustment;
-    uint64_t remote_ack_ptr_addr = get_noc_addr(sender_noc_x, sender_noc_y, (uint32_t)pages_acked_ptr, noc);
+    uint64_t remote_ack_ptr_addr = get_noc_addr(sender_noc_x, sender_noc_y, (uintptr_t)pages_acked_ptr, noc);
     noc_fast_atomic_increment<nm>(
         noc,
         cmd_buf,
